+++ android/src/main/java/com/idlefish/flutterboost/FlutterBoostPlugin.java
@@ -271,6 +271,9 @@ public class FlutterBoostPlugin implements FlutterPlugin, NativeRouterApi, Activ
             CommonParams params = new CommonParams();
             channel.onForeground(params, reply -> {
             });
+
+            // The scheduling frames are resumed when [onForeground] is called.
+            changeFlutterAppLifecycle(FLUTTER_APP_STATE_RESUMED);
         } else {
             throw new RuntimeException("FlutterBoostPlugin might *NOT* have attached to engine yet!");
         }
@@ -283,6 +286,9 @@ public class FlutterBoostPlugin implements FlutterPlugin, NativeRouterApi, Activ
             CommonParams params = new CommonParams();
             channel.onBackground(params, reply -> {
             });
+
+            // The scheduling frames are paused when [onBackground] is called.
+            changeFlutterAppLifecycle(FLUTTER_APP_STATE_PAUSED);
         } else {
             throw new RuntimeException("FlutterBoostPlugin might *NOT* have attached to engine yet!");
         }